ENH: Pass path to R as an option to the cmake call
Turns out that the travis/R, which presumably copies the
CRAN setup, has a sneaky trick to ensure that R is never
called without a full path. CMake appears to use a search
for R, presumably via the path, to locate R and picks up
the fake version that travis/cran places to trap this
behaviour.