WIP: Add support for Windows
Add support for building and testing on Windows OS
in a consistent manner to Linux and macOS.
Repository URL and tag are specified in a shared
configuration file (SITK_CONFIG) which is used
by both the configure and configure.win scripts.
Testing is performed in CircleCI as it offers a
more powerful build machine under the free plan.

Co-authored-by: Artur-man <artur-man@hotmail.com>